AI/requirements.txt
google-genai
python-dotenv==1.0.0
pytesseract==0.3.10
pillow==11.0.0
pdfplumber
fastapi
uvicorn
python-multipart
httpx
pymongo
qdrant-client
